requirements.txt
streamlit==1.28.0
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.0
Pillow==10.0.0
opencv-python-headless==4.8.1.78